Refactor Arrhenius components
