+++ components/ArrheniusGraph.tsx
@@ -1,117 +1,49 @@
 "use client";
 
-import Link from "next/link";
-import { BlockMath, InlineMath } from "react-katex";
-import "katex/dist/katex.min.css";
-import ArrheniusGraph from "@/components/ArrheniusGraph";
+import { useMemo } from "react";
+import {
+  LineChart,
+  Line,
+  XAxis,
+  YAxis,
+  CartesianGrid,
+  Tooltip,
+  ResponsiveContainer,
+} from "recharts";
 
-export default function ArrheniusArticle() {
+interface ArrheniusGraphProps {
+  Ea?: number; // Activation energy (eV)
+  useTemp?: number; // Use temperature in °C
+}
+
+const k = 8.617e-5; // Boltzmann constant in eV/K
+
+export default function ArrheniusGraph({ Ea = 0.8, useTemp = 55 }: ArrheniusGraphProps) {
+  const data = useMemo(() => {
+    const points = [] as { temperature: number; AF: number }[];
+    const useTempK = useTemp + 273.15;
+    for (let stress = useTemp; stress <= 150; stress += 5) {
+      const stressK = stress + 273.15;
+      const AF = Math.exp((Ea / k) * (1 / useTempK - 1 / stressK));
+      points.push({ temperature: stress, AF: parseFloat(AF.toFixed(2)) });
+    }
+    return points;
+  }, [Ea, useTemp]);
+
   return (
-    <main className="max-w-3xl mx-auto px-6 py-10">
-      <h1 className="text-4xl font-bold mb-4">Understanding Arrhenius for Reliability</h1>
-      <p className="text-lg text-gray-700 mb-6">
-        Reliability engineering is all about <strong>predicting how long a product will last under specific conditions</strong>.
-        One of the most fundamental tools for this prediction is the <strong>Arrhenius equation</strong>, which models the effect
-        of temperature on reaction rates—and by extension, on failure mechanisms in materials and systems.
-      </p>
-
-      <h2 className="text-2xl font-semibold mt-8 mb-4">The Arrhenius Equation: A Foundation of Reliability Modeling</h2>
-      <p>The Arrhenius equation mathematically relates the rate of a chemical reaction (or failure process) to temperature:</p>
-      <BlockMath math={"t_f = A \\cdot e^{-\\frac{E_a}{kT}}"} />
-
-      <p>Where:</p>
-      <ul className="list-disc list-inside">
-        <li><strong>t<sub>f</sub></strong>: Time to failure</li>
-        <li><strong>A</strong>: Pre-exponential factor (a constant for the material or process)</li>
-        <li><strong>E<sub>a</sub></strong>: Activation energy (the energy barrier for the reaction)</li>
-        <li><strong>k</strong>: Boltzmann constant (8.617 × 10⁻⁵ eV/K)</li>
-        <li><strong>T</strong>: Absolute temperature in Kelvin</li>
-      </ul>
-
-      <p>This equation shows that <strong>as temperature increases, time to failure decreases exponentially</strong>—a critical insight for accelerated testing.</p>
-
-      <h2 className="text-2xl font-semibold mt-8 mb-4">Graph: Acceleration Factor vs. Temperature</h2>
-      <p>The graph below illustrates how the acceleration factor increases with temperature for a typical activation energy:</p>
-
-      <ArrheniusGraph />
-
-      <h2 className="text-2xl font-semibold mt-8 mb-4">Why Activation Energy Matters</h2>
-      <p>
-        Activation energy <InlineMath math={"E_a"} /> represents the minimum energy required for a reaction to occur.
-        In reliability, this translates to the energy needed for a failure mechanism (e.g., electromigration, thermal breakdown) to manifest.
-      </p>
-
-      <p>Different materials and failure modes have different <InlineMath math={"E_a"} /> values:</p>
-      <ul className="list-disc list-inside">
-        <li>For metals in microelectronics (e.g., aluminum, copper), <InlineMath math={"E_a"} /> typically ranges from 0.5 to 1.0 eV.</li>
-        <li>For polymers in automotive components, values can vary widely based on additives, fillers, and degradation modes (thermal oxidation, hydrolysis).</li>
-      </ul>
-
-      <h2 className="text-2xl font-semibold mt-8 mb-4">Acceleration Factor: Predicting Life from Tests</h2>
-      <p>
-        One of the most powerful uses of Arrhenius is calculating the <strong>Acceleration Factor (AF)</strong>—how much faster a failure occurs at an elevated test temperature compared to normal use:
-      </p>
-      <BlockMath math={"AF = \\exp \\left( \\frac{E_a}{k} \\left( \\frac{1}{T_{use}} - \\frac{1}{T_{test}} \\right) \\right)"} />
-
-      <p>For example, an automotive control module might undergo a 1000-hour test at 125°C. Using Arrhenius, engineers can estimate that this is equivalent to 10 years at 55°C in the field. This capability is essential for industries like:</p>
-      <ul className="list-disc list-inside">
-        <li><strong>Automotive:</strong> Validating ECUs, sensors, and connectors to withstand years of harsh thermal exposure.</li>
-        <li><strong>Microelectronics:</strong> Testing chips and packages under accelerated thermal and voltage stresses to predict long-term reliability.</li>
-      </ul>
-
-      <h2 className="text-2xl font-semibold mt-8 mb-4">Real-World Applications</h2>
-      <h3 className="text-xl font-semibold mt-6 mb-2">Automotive Industry: Simulating a Decade in Weeks</h3>
-      <p>
-        In the <strong>Robustness Validation</strong> process for automotive electronic modules, the Arrhenius model underpins burn-in tests and mission profile validations.
-        By applying realistic environmental stressors (e.g., -40°C to +125°C temperature ranges, vibration, humidity), engineers ensure that control units, sensors, and actuators can survive 15 years of vehicle life.
-        The <a href="https://www.zvei.org/fileadmin/user_upload/Presse_und_Medien/Publikationen/2013/Oktober/Handbook_for_Robustness_Validation_of_Automotive_Electrical/Robust-Validation-EE-Modules.pdf" 
-        target="_blank" 
-        rel="noopener noreferrer"
-        className="text-blue-600 hover:underline">ZVEI Handbook</a> provides structured methodologies for integrating Arrhenius-based models into testing protocols.
-      </p>
-      <p>
-        Curious how simulate temperature affects? Check out our{" "}
-        <Link href="/app/tools/Arrhenius/page.tsx" 
-        className="text-blue-600 hover:underline"
-        >Arrhenius Calculator</Link> for quick estimates.
-      </p>
-
-      <h3 className="text-xl font-semibold mt-6 mb-2">Microelectronics: Managing Thermal Risks</h3>
-      <p>  
-        In <strong>microelectronics</strong>, the Arrhenius equation helps predict failure mechanisms like electromigration and dielectric breakdown.
-        For example, a chip operating at 85°C may fail in 10 years, but testing at 150°C can simulate this lifetime in just a few months.
-        The <a href="https://ntrs.nasa.gov/citations/20230004376" 
-        target="_blank" 
-        rel="noopener noreferrer"
-        className="text-blue-600 hover:underline"
-        >NASA Physics of Failure Handbook</a> emphasizes the importance of understanding the underlying failure physics—not just applying the equation blindly.
-      </p>
-      <p>
-        Want to estimate your own burn-in conditions to ensure quality? Try our{" "}
-        <Link href="/tools/BurnInWizard" className="text-blue-600 hover:underline">Burn-In Wizard</Link> to simulate equivalent field life from accelerated test plans.
-      </p>
-      
-
-      <h2 className="text-2xl font-semibold mt-8 mb-4">Limitations and Cautions</h2>
-      <ul className="list-disc list-inside">
-        <li><strong>Not all failures are thermally activated:</strong> Mechanical fatigue, corrosion, and radiation may require different models.</li>
-        <li><strong>Test temperatures must be realistic:</strong> Exceeding material limits can introduce failure mechanisms (e.g., melting solder, dielectric breakdown) that wouldn't occur in the field.</li>
-        <li><strong>Assumptions of constant <InlineMath math={"E_a"} />:</strong> May not hold across different materials, temperatures, or combined stresses (e.g., temperature and humidity in the Peck model).</li>
-      </ul>
-
-      <h2 className="text-2xl font-semibold mt-8 mb-4">Conclusion: A Tool, Not a Crystal Ball</h2>
-      <p>
-        The Arrhenius equation remains a cornerstone of accelerated life testing and physics-of-failure reliability engineering.
-        By understanding the materials, the failure modes, and the operating environment, engineers can use Arrhenius-based models to predict, validate, and improve product reliability—whether it’s a car’s control module surviving 15 years on the road or a microchip enduring a decade of power cycling.
-      </p>
-
-      <p>
-        Ready to try it yourself? Visit our{" "}
-        <Link href="/tools/Arrhenius" className="text-blue-600 hover:underline">Arrhenius Calculator</Link>{" "}
-        and{" "}
-        <Link href="/tools/BurnInWizard" className="text-blue-600 hover:underline">Burn-In Wizard</Link>{" "}
-        to simplify your reliability test planning today.
-      </p>
-    </main>
+    <div className="h-64">
+      <ResponsiveContainer width="100%" height="100%">
+        <LineChart data={data}>
+          <CartesianGrid strokeDasharray="3 3" />
+          <XAxis
+            dataKey="temperature"
+            label={{ value: "Stress Temp (°C)", position: "insideBottom", offset: -5 }}
+          />
+          <YAxis label={{ value: "AF", angle: -90, position: "insideLeft" }} />
+          <Tooltip formatter={(value: number) => value.toFixed(2)} />
+          <Line type="monotone" dataKey="AF" stroke="#2563EB" strokeWidth={2} dot={false} />
+        </LineChart>
+      </ResponsiveContainer>
+    </div>
   );
 }
